Extract discover_repos to core, make diff_similarity pub
Move discover_repos() from git-branch-tidy to git-tidy-core::discovery
so all future tools can reuse it. git-branch-tidy now re-exports from
core. Make diff_similarity() pub for reuse by stash-tidy.
